requirements.txt
pandas
numpy
requests
scikit-learn
yfinance
gspread
oauth2client
python-dotenv
